+++ lib/src/emoji.dart
@@ -1,7 +1,11 @@
 class Emoji {
   static const String shout = '😱';
-  static const String severe = '🚫';
+  static const String severe = '🛑';
+  static const String error = '🚫';
+  // static const String warning = '⚠️';
   static const String warning = '🚧';
   static const String info = '📗';
+  // static const String config = '⚙️';
+  static const String config = '🔧';
   static const String debug = '🐛';
 }