+++ lib/src/hrk_level.dart
@@ -1,10 +1,12 @@
+// ignore_for_file: constant_identifier_names
+
 import 'package:logging/logging.dart';
 
 class HrkLevel {
-  // ignore: constant_identifier_names
   static const Level DEBUG = Level('DEBUG', 600);
 
-  // ignore: constant_identifier_names
+  static const Level ERROR = Level('ERROR', 950);
+
   static const List<Level> LEVELS = [
     Level.ALL,
     Level.FINEST,
@@ -14,6 +16,7 @@ class HrkLevel {
     Level.CONFIG,
     Level.INFO,
     Level.WARNING,
+    HrkLevel.ERROR,
     Level.SEVERE,
     Level.SHOUT,
     Level.OFF
@@ -23,4 +26,7 @@ class HrkLevel {
 extension HrkLogger on Logger {
   void debug(Object? message, [Object? error, StackTrace? stackTrace]) =>
       log(HrkLevel.DEBUG, message, error, stackTrace);
+
+  void error(Object? message, [Object? error, StackTrace? stackTrace]) =>
+      log(HrkLevel.ERROR, message, error, stackTrace);
 }