+++ lib/src/hrk_logging.dart
@@ -6,11 +6,11 @@ import 'package:logging/logging.dart';
 import 'ansi_color.dart';
 import 'emoji.dart';
 import 'globals.dart';
-import 'level_debug.dart';
+import 'hrk_level.dart';
 
 export 'ansi_color.dart';
 export 'emoji.dart';
-export 'level_debug.dart';
+export 'hrk_level.dart';
 
 /// To maintain idempotency of the configureLogging()
 StreamSubscription<LogRecord>? rootLoggerSubscription;
@@ -23,6 +23,7 @@ void configureHrkLogging() {
   hierarchicalLoggingEnabled = true;
   Level level = kDebugMode ? Level.ALL : Level.INFO;
   Logger.root.level = dartOrFlutterTest ? Level.OFF : level;
+
   rootLoggerSubscription = Logger.root.onRecord.listen((record) {
     late final String color;
     late final String emoji;
@@ -30,9 +31,12 @@ void configureHrkLogging() {
     late final String resetColor;
     if (record.level == Level.SHOUT) {
       emoji = Emoji.shout;
-      color = AnsiColor.red;
+      color = AnsiColor.brightRed;
     } else if (record.level == Level.SEVERE) {
       emoji = Emoji.severe;
+      color = AnsiColor.brightRed;
+    } else if (record.level == HrkLevel.ERROR) {
+      emoji = Emoji.error;
       color = AnsiColor.red;
     } else if (record.level == Level.WARNING) {
       emoji = Emoji.warning;
@@ -40,6 +44,9 @@ void configureHrkLogging() {
     } else if (record.level == Level.INFO) {
       emoji = Emoji.info;
       color = AnsiColor.green;
+    } else if (record.level == Level.CONFIG) {
+      emoji = Emoji.config;
+      color = AnsiColor.magenta;
     } else if (record.level == HrkLevel.DEBUG) {
       emoji = Emoji.debug;
       color = AnsiColor.blue;
@@ -54,7 +61,9 @@ void configureHrkLogging() {
       '$color'
       '$emoji'
       '$emojiSpacer'
-      '${record.level.name.capitalize()}: ${record.time}: ${record.message}'
+      '${record.level.name.capitalize()}: '
+      '${record.time}: '
+      '${record.message}'
       '$resetColor',
     );
   });