+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.0
+
+- Add HrkLevel.ERROR in between Level.WARNING and Level.SEVERE
+- Assign 🚫 emoji to Level.ERROR
+- Assign 🛑 emoji to Level.SEVERE
+- Assign Brigh Red color to Level.SEVERE and Level.SHOUT
+- Assign 🔧 emoji to CONFIG
+- Assign Magenta color to CONFIG
+
 ## 0.1.0-pre.3
 
 - Integrate hrk_analysis